+++ parselcli/prompt.py
@@ -4,22 +4,24 @@ import shlex
 import webbrowser
 
 import re
+from collections import OrderedDict
 from inspect import signature
+from tempfile import NamedTemporaryFile
+from typing import List
 
 import click
 import requests
-from click import echo, BadOptionUsage, NoSuchOption
+from click import echo, BadOptionUsage, NoSuchOption, Tuple
 from parsel import Selector
 from prompt_toolkit import prompt
 from prompt_toolkit.auto_suggest import AutoSuggestFromHistory
 from prompt_toolkit.history import FileHistory
 from requests import Response
 
-from parselcli.app import HISTORY
 from parselcli.completer import MiddleWordCompleter
 from parselcli.embed import embed_auto
 from parselcli.parse import ToggleInputOptionParser
-from parselcli.processors import Strip, First, UrlJoin, Join, Collapse
+from parselcli.processors import Strip, First, AbsoluteUrl, Join, Collapse, Len
 from parselcli.completer import XPATH_COMPLETION, CSS_COMPLETION, BASE_COMPLETION
 
 HELP = {
@@ -30,9 +32,11 @@ HELP = {
     'first': "take first element when there's only one",
     'help': 'show help',
     'join': 'join results into one',
+    'len': 'return length of results',
     'collapse': 'collapse lists when only 1 element',
     'strip': 'strip every element of trailing and leading spaces',
-    'view': 'open current file in browser tab',
+    'open': 'open current url in browser tab',
+    'view': 'open current html in browser tab',
     'css': 'switch to css selectors',
     'xpath': 'switch to xpath selectors',
 }
@@ -76,14 +80,16 @@ class Prompter:
     """
     Prompt Toolkit container for all interpreter functions
     """
+    available_flags = ['strip', 'collapse', 'absolute', 'join', 'first', 'len']
 
     def __init__(
             self,
             selector: Selector,
             response: Response = None,
-            preferred_embed_shell=None,
+            preferred_embed_shell: str = None,
             start_in_css=False,
-            flags=('collapse',)):
+            history_file=None,
+            flags: List = None):
         """
         :param selector:
         :param response:
@@ -91,6 +97,11 @@ class Prompter:
         :param start_in_css: whether to start in css mode instead of xpath
         :param flags: default flags to enable
         """
+        self._option_parser = None
+        self._flags = None
+        self._commands = None
+
+        self.history_file = history_file
         self.sel = selector
         self.response = response
         self.processors = []
@@ -103,30 +114,32 @@ class Prompter:
             self.completer = self.completer_css
 
         # setup interpreter flags and commands
-        flags = flags or []
         for flag in flags:
             self.enable_flag(flag)
-        # self.commands = self._get_commands()
-        self._option_parser = None
-        self._flags = None
-        self._commands = None
 
     @classmethod
-    def from_response(cls, response, preferred_embed_shell=None, start_in_css=False, flags=None):
+    def from_response(cls, response, preferred_embed_shell=None, start_in_css=False, flags=None, history_file=None):
         return cls(
             Selector(response.text),
             response=response,
             preferred_embed_shell=preferred_embed_shell,
             start_in_css=start_in_css,
+            history_file=history_file,
             flags=flags)
 
     @classmethod
-    def from_file(cls, file, *args, **kwargs):
+    def from_file(cls, file, preferred_embed_shell=None, start_in_css=False, flags=None, history_file=None):
         response = Response()
         response._content = file.read().encode('utf8')
         response.status_code = 200
         response.url = pathlib.Path(os.path.abspath(file.name)).as_uri()
-        return cls(Selector(response.text), response=response, *args, **kwargs)
+        return cls(
+            Selector(response.text),
+            response=response,
+            preferred_embed_shell=preferred_embed_shell,
+            start_in_css=start_in_css,
+            flags=flags,
+            history_file=history_file)
 
     def _create_completers(self, selector):
         """Initiated auto completers based on current selector"""
@@ -148,6 +161,7 @@ class Prompter:
                 'help': self.cmd_help,
                 'debug': self.cmd_debug,
                 'embed': self.cmd_embed,
+                'open': self.cmd_open,
                 'view': self.cmd_view,
                 'fetch': self.cmd_fetch,
                 'css': self.cmd_switch_to_css,
@@ -162,8 +176,9 @@ class Prompter:
             self._flags['strip'] = Strip()
             self._flags['first'] = First()
             self._flags['collapse'] = Collapse()
-            self._flags['absolute'] = UrlJoin(self.response.url)
+            self._flags['absolute'] = AbsoluteUrl(self.response.url)
             self._flags['join'] = Join()
+            self._flags['len'] = Len()
         return self._flags
 
     @property
@@ -186,8 +201,13 @@ class Prompter:
         self.sel = Selector(text=self.response.text)
         self._create_completers(self.sel)
 
-    def cmd_view(self):
+    def cmd_open(self):
         webbrowser.open_new_tab(self.response.url)
+
+    def cmd_view(self):
+        with NamedTemporaryFile('w', encoding=self.response.encoding, delete=False, suffix='.html') as file:
+            file.write(self.response.text)
+        webbrowser.open_new_tab(f'file://{file.name}')
 
     def cmd_help(self):
         print('available commands (use -command):')
@@ -226,50 +246,56 @@ class Prompter:
         if past > len(self.processors):
             print('disabled flag: {}'.format(flag))
 
-    def process(self, data):
+    def process(self, data, processors=None):
         """Process data through enabled flag processors"""
+        if not processors:
+            processors = self.processors
         try:
-            for processor in self.processors:
+            for processor in processors:
                 data = processor(data)
         except Exception as e:
             print(e)
         return data
 
-    def get_xpath(self, text):
+    def get_xpath(self, text, processors=None):
         """Tries to extract xpath from a selector"""
         try:
-            return self.process(self.sel.xpath(text).extract())
+            return self.process(self.sel.xpath(text).extract(), processors=processors)
         except Exception as f:
-            echo('E: {}'.format(f))
-            return self.process([])
+            echo(f'E:"{text}": {f}')
+            return self.process([], processors=processors)
 
-    def get_css(self, text):
+    def get_css(self, text, processors=None):
         """Tries to extract css from a selector"""
         try:
-            return self.process(self.sel.css(text).extract())
+            return self.process(self.sel.css(text).extract(), processors=processors)
         except Exception as f:
-            echo('E: {}'.format(f))
-            return self.process([])
+            echo(f'E:"{text}": {f}')
+            return self.process([], processors=processors)
 
     def parse_commands(self, text):
+        """
+        Parses commands and flags from a string
+        :returns remaining_text, found_commands, found_flags_to_enable, found_flags_to_disable
+        """
         cmds = shlex.split(text)
-        try:
-            parsed = self.option_parser.parse_args(cmds)[0].items()
-        except (BadOptionUsage, NoSuchOption) as e:
-            print(e)
-            parsed = []
-        for k, v in parsed:
+        parsed, _, _ = self.option_parser.parse_args(cmds)
+        commands = OrderedDict()
+        flags_enable = []
+        flags_disable = []
+        for k, v in parsed.items():
+            text = text.replace(f'{k} {v or ""}'.strip(), '')
             if not isinstance(v, (list, tuple)):
                 v = [v]
             notation, k = [x for x in re.split('([+-])', k) if x]
             if k in self.commands:
-                self.commands[k](*v)
-                continue
+                commands[k] = v
             if k in self.flags:
                 if notation == '+':
-                    self.enable_flag(k)
+                    flags_enable.append(k)
                 else:
-                    self.disable_flag(k)
+                    flags_disable.append(k)
+        return text.strip(), commands, flags_enable, flags_disable
 
     def cmd_switch_to_css(self):
         print('switched to css')
@@ -281,7 +307,7 @@ class Prompter:
 
     def loop_prompt(self, start_in_embed=False):
         """main loop prompt"""
-        prompt_history = FileHistory(HISTORY)
+        prompt_history = FileHistory(self.history_file)
         while True:
             if start_in_embed:
                 self.cmd_embed()
@@ -292,19 +318,37 @@ class Prompter:
                           completer=self.completer)
             if not text:
                 continue
+            processors = self.processors
             # check flags and commands
-            if re.match(r'[+-]\w+', text):
-                self.parse_commands(text)
-                continue
-            if text == 'e':
-                self.cmd_embed()
+            if re.findall(r'[+-]\w+', text):
+                try:
+                    text, commands, flags_enable, flags_disable = self.parse_commands(text)
+                except (BadOptionUsage, NoSuchOption) as e:
+                    echo(e)
+                    continue
+                # if any commands are found execute first one
+                for cmd_name, value in commands.items():
+                    self.commands[cmd_name](*value)
+                    continue
+                # if there is no text left enable and disable commands
+                if not text.strip():
+                    for flag in flags_enable:
+                        self.enable_flag(flag)
+                    for flag in flags_disable:
+                        self.disable_flag(flag)
+                    continue
+                # enable temporary flags
+                processors = [self.flags[flag] for flag in flags_enable]
             if self.completer is self.completer_css:
-                value = self.get_css(text)
+                value = self.get_css(text, processors=processors)
             else:
-                value = self.get_xpath(text)
-            value_len = len(''.join(value))
-            if value_len < 1000:
-                echo(value)
-            else:
-                if click.confirm(f'very big output {value_len}, print?'):
+                value = self.get_xpath(text, processors=processors)
+            try:
+                value_len = len(''.join(value))
+                if value_len < 2000:
                     echo(value)
+                else:
+                    if click.confirm(f'very big output {value_len}, print?'):
+                        echo(value)
+            except TypeError:
+                echo(value)